Add files via upload
The project is about to control street light. It senses the external light and then accordingly switches ON or OFF the street lights and vehicle detection. This project help to decrease the wastage of electricity by controlling the working of street light system that attributes to a good amount of electricity bills in our nation